changed type of ids to int64

+++ internal/domain/models/application_phase.go
@@ -3,10 +3,10 @@ package models
 import "time"
 
 type ApplicationPhase struct {
-	ID            uint      `json:"id" db:"id"`
+	ID            int64     `json:"id" db:"id"`
 	Name          string    `json:"name" db:"name"`
 	Date          time.Time `json:"date" db:"date"`
 	Created       time.Time `json:"created" db:"created"`
 	Notes         string    `json:"notes" db:"notes"`
-	ApplicationID uint      `json:"application_id" db:"application_id"`
+	ApplicationID int64     `json:"application_id" db:"application_id"`
 }
