+++ internal/domain/models/user.go
@@ -1,7 +1,7 @@
 package models
 
 type User struct {
-	ID       uint   `json:"id" db:"id"`
+	ID       int64  `json:"id" db:"id"`
 	Password string `json:"password" db:"password"`
 	Email    string `json:"email" db:"email"`
 	Name     string `json:"name" db:"name"`
